requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.10
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
confection==0.1.0
cymem==2.0.7
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
joblib==1.3.1
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
murmurhash==1.0.9
nltk==3.8.1
numpy==1.25.1
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==10.0.0
preshed==3.0.8
psycopg2==2.9.9
pycparser==2.21
pydantic==1.10.11
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
thinc==8.1.10
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.6
urllib3==2.0.4
wasabi==1.1.2
wsproto==1.2.0
